requirements.txt
requests>=2.28.0
psutil>=5.9.0
pathlib2>=2.3.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Vector Database & Semantic Search
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Computer Vision
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=9.0.0
pyautogui>=0.9.54
easyocr>=1.7.0
ultralytics>=8.0.0

# Sentiment Analysis & NLP
textblob>=0.17.0
vaderSentiment>=3.3.0
transformers>=4.30.0
torch>=2.0.0

# Calendar & Scheduling
schedule>=1.2.0
icalendar>=5.0.0

# Encryption & Security
cryptography>=41.0.0
bcrypt>=4.0.0

# ML & AI
networkx>=3.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0